Add wrappers, port tests and docs
