+++ crates/bevy_tasks/src/lib.rs
@@ -9,13 +9,12 @@ pub use task::Task;
 
 #[cfg(not(target_arch = "wasm32"))]
 mod task_pool_non_wasm;
-#[cfg(not(target_arch = "wasm32"))]
-pub use task_pool_non_wasm::{Scope, TaskPool, TaskPoolBuilder};
+#[cfg(target_arch = "wasm32")]
+mod task_pool_wasm;
 
-#[cfg(target_arch = "wasm32")]
-mod single_threaded_task_pool;
-#[cfg(target_arch = "wasm32")]
-pub use single_threaded_task_pool::{Scope, TaskPool, TaskPoolBuilder};
+mod task_pool;
+#[cfg(not(target_arch = "wasm32"))]
+pub use task_pool::{Scope, TaskPool, TaskPoolBuilder};
 
 mod usages;
 pub use usages::{AsyncComputeTaskPool, ComputeTaskPool, IoTaskPool};
