+++ crates/bevy_tasks/src/task_pool_wasm.rs
@@ -1,3 +1,5 @@
+//! TaskPool implementation for wasm platform.
+
 use std::{
     future::Future,
     marker::PhantomData,
@@ -5,12 +7,12 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-/// Used to create a TaskPool
+/// Used to create a PlatformTaskPool
 #[derive(Debug, Default, Clone)]
-pub struct TaskPoolBuilder {}
+pub struct PlatformTaskPoolBuilder {}
 
-impl TaskPoolBuilder {
-    /// Creates a new TaskPoolBuilder instance
+impl PlatformTaskPoolBuilder {
+    /// Creates a new PlatformTaskPoolBuilder instance
     pub fn new() -> Self {
         Self::default()
     }
@@ -30,21 +32,21 @@ impl TaskPoolBuilder {
         self
     }
 
-    /// Creates a new [`TaskPool`]
-    pub fn build(self) -> TaskPool {
-        TaskPool::new_internal()
+    /// Creates a new [`PlatformTaskPool`]
+    pub fn build(self) -> PlatformTaskPool {
+        PlatformTaskPool::new_internal()
     }
 }
 
 /// A thread pool for executing tasks. Tasks are futures that are being automatically driven by
 /// the pool on threads owned by the pool. In this case - main thread only.
 #[derive(Debug, Default, Clone)]
-pub struct TaskPool {}
+pub struct PlatformTaskPool {}
 
-impl TaskPool {
-    /// Create a `TaskPool` with the default configuration.
+impl PlatformTaskPool {
+    /// Create a `PlatformTaskPool` with the default configuration.
     pub fn new() -> Self {
-        TaskPoolBuilder::new().build()
+        PlatformTaskPoolBuilder::new().build()
     }
 
     #[allow(unused_variables)]
@@ -64,7 +66,7 @@ impl TaskPool {
     /// This is similar to `rayon::scope` and `crossbeam::scope`
     pub fn scope<'env, F, T>(&self, f: F) -> Vec<T>
     where
-        F: for<'scope> FnOnce(&'env mut Scope<'scope, 'env, T>),
+        F: for<'scope> FnOnce(&'env mut PlatformScope<'scope, 'env, T>),
         T: Send + 'static,
     {
         let executor = &async_executor::LocalExecutor::new();
@@ -74,14 +76,14 @@ impl TaskPool {
         let results: Mutex<Vec<Arc<Mutex<Option<T>>>>> = Mutex::new(Vec::new());
         let results: &'env Mutex<Vec<Arc<Mutex<Option<T>>>>> = unsafe { mem::transmute(&results) };
 
-        let mut scope = Scope {
+        let mut scope = PlatformScope {
             executor,
             results,
             scope: PhantomData,
             env: PhantomData,
         };
 
-        let scope_ref: &'env mut Scope<'_, 'env, T> = unsafe { mem::transmute(&mut scope) };
+        let scope_ref: &'env mut PlatformScope<'_, 'env, T> = unsafe { mem::transmute(&mut scope) };
 
         f(scope_ref);
 
@@ -131,37 +133,37 @@ impl FakeTask {
     pub fn detach(self) {}
 }
 
-/// A `TaskPool` scope for running one or more non-`'static` futures.
+/// A `PlatformTaskPool` scope for running one or more non-`'static` futures.
 ///
-/// For more information, see [`TaskPool::scope`].
+/// For more information, see [`PlatformTaskPool::scope`].
 #[derive(Debug)]
-pub struct Scope<'scope, 'env: 'scope, T> {
+pub struct PlatformScope<'scope, 'env: 'scope, T> {
     executor: &'env async_executor::LocalExecutor<'env>,
     // Vector to gather results of all futures spawned during scope run
     results: &'env Mutex<Vec<Arc<Mutex<Option<T>>>>>,
 
-    // make `Scope` invariant over 'scope and 'env
+    // make `PlatformScope` invariant over 'scope and 'env
     scope: PhantomData<&'scope mut &'scope ()>,
     env: PhantomData<&'env mut &'env ()>,
 }
 
-impl<'scope, 'env, T: Send + 'env> Scope<'scope, 'env, T> {
+impl<'scope, 'env, T: Send + 'env> PlatformScope<'scope, 'env, T> {
     /// Spawns a scoped future onto the thread-local executor. The scope *must* outlive
     /// the provided future. The results of the future will be returned as a part of
-    /// [`TaskPool::scope`]'s return value.
+    /// [`PlatformTaskPool::scope`]'s return value.
     ///
-    /// On the single threaded task pool, it just calls [`Scope::spawn_local`].
+    /// On the single threaded task pool, it just calls [`PlatformScope::spawn_local`].
     ///
-    /// For more information, see [`TaskPool::scope`].
+    /// For more information, see [`PlatformTaskPool::scope`].
     pub fn spawn<Fut: Future<Output = T> + 'env>(&self, f: Fut) {
         self.spawn_on_scope(f);
     }
 
     /// Spawns a scoped future that runs on the thread the scope called from. The
     /// scope *must* outlive the provided future. The results of the future will be
-    /// returned as a part of [`TaskPool::scope`]'s return value.
+    /// returned as a part of [`PlatformTaskPool::scope`]'s return value.
     ///
-    /// For more information, see [`TaskPool::scope`].
+    /// For more information, see [`PlatformTaskPool::scope`].
     pub fn spawn_on_scope<Fut: Future<Output = T> + 'env>(&self, f: Fut) {
         let result = Arc::new(Mutex::new(None));
         self.results.lock().unwrap().push(result.clone());
